+++ src/builder/b-pollen.ts
@@ -2,8 +2,9 @@ import { HomeAssistant } from 'custom-card-helpers/dist';
 import { iPollen } from '../utils/config-schema';
 import { getEntityRawValue } from '../utils/entity';
 import { iPollenData, renderPollen } from '../templates/t-pollen';
+import { iTerms } from '../base/lovelace-base';
 
-const buildPollen = (hass: HomeAssistant, pollen: iPollen) => {
+const buildPollen = (hass: HomeAssistant, pollen: iPollen, terms: iTerms) => {
   const allItems: iPollenData[] = [];
   if (Array.isArray(pollen.entities) && pollen.entities.length > 0) {
     pollen.entities.forEach((item) => {
@@ -17,7 +18,7 @@ const buildPollen = (hass: HomeAssistant, pollen: iPollen) => {
     });
   }
 
-  return renderPollen(allItems, pollen.min, pollen.max);
+  return renderPollen(allItems, pollen.min, pollen.max, terms.words);
 };
 
 export default buildPollen;